change unsigned to unsigned long long for overflow

+++ c++/a_generate_files/fct_generate_files.hpp
@@ -11,10 +11,10 @@ bool read_ini_file(const char* ini_filename, std::string& path_to_gz,
 	std::string& path_to_output, unsigned& nb_ngrams, unsigned& min_year_defined);
 	
 void treat_line(FILE* output, std::string& ngram, std::string& precedent_ngram, unsigned& somme_year,
-	unsigned& somme_nb_match, unsigned& somme_nb_volume, float& mean_pondere_match,
-	float& mean_pondere_volume,	unsigned& year_max,	unsigned& year_min,
+	unsigned long long& somme_nb_match, unsigned long long& somme_nb_volume, unsigned long long& mean_pondere_match,
+	unsigned long long& mean_pondere_volume,	unsigned& year_max,	unsigned& year_min,
 	unsigned& nb_match_max,	unsigned& nb_match_min,	unsigned& nb_volume_max,
-	unsigned& nb_volume_min, unsigned year, unsigned nb_match, unsigned nb_volume);
+	unsigned& nb_volume_min, unsigned long long year, unsigned long long nb_match, unsigned long long nb_volume);
 	
 bool file_not_entirely_read(std::stringstream& token);
