+++ c++/a_generate_files/fct_valid_lines.hpp
@@ -12,7 +12,7 @@ void collect_words_tags(std::string ngram, std::vector<std::string>& words_tags)
 bool valid_ngram(std::string ngram, std::vector<std::string>& forbidden_characters, std::vector<std::string>& accepted_tags, unsigned nb_ngram);
 
 bool valid_line(std::string line, std::string& ngram, 
-	unsigned& year, unsigned& nb_match, unsigned& nb_volume, 
+	unsigned long long& year, unsigned long long& nb_match, unsigned long long& nb_volume, 
 	std::vector<std::string>& forbidden_characters, 
 	std::vector<std::string>& accepted_tags, 
 	unsigned nb_ngram, unsigned min_year_defined);
